made dynamic

+++ appointment/src/app/app.component.ts
@@ -1,8 +1,7 @@
-import { Component, OnInit,Inject } from '@angular/core';
- import {Router} from "@angular/router";
+import { Component, OnInit, Inject } from '@angular/core';
+import { Router, ActivatedRoute } from "@angular/router";
 import { Appointments } from './model/Appointment.model'
 import { ApiService } from './service/api.service';
-
 @Component({
   selector: 'list-component',
   templateUrl: './app.component.html',
@@ -10,38 +9,45 @@ import { ApiService } from './service/api.service';
 })
 export class ListAppointmentComponent implements OnInit {
 
-  appointments : Appointments[];
-  constructor (
-    private apiservice : ApiService,private router: Router,) { }
+  appointments: Appointments[];
+  public memberId: any;
+  constructor(
+    private apiservice: ApiService, private router: Router, private route: ActivatedRoute) { }
 
   ngOnInit() {
     //get Users List
-    this.apiservice.getAppointments().subscribe(data=>{
+    this.memberId = this.route.snapshot.paramMap.get('memberId');
+    this.apiservice.getAppointments(this.memberId).subscribe(data => {
       console.log(data);
-     this.appointments=data;
+      this.appointments = data;
     })
   }
 
-  addUser() : void {
+  addUser(): void {
     this.router.navigate(['add']);
   }
 
-  editAppointment(id) {
-   this.router.navigate(['/edit', id]);
+  editAppointment(id, memberId) {
+    this.router.navigate(['edit', id, memberId]);
+
   }
 
-  cancelAppointemnt(id,memberId) {
-    console.log(id)
-    this.apiservice.deleteAppointment(id,memberId).subscribe(data=>{
-      if (data.statusCode!=200){
-        alert(data.Message);
-      }else{
-        alert("Rescheduled Successfully");
-      }
-      this.apiservice.getAppointments().subscribe(data=>{
-        console.log(data);
-       this.appointments=data;
+  cancelAppointemnt(id, memberId) {
+    if (confirm('Are you sure you want to cancel the appointment?')) {
+      this.apiservice.deleteAppointment(id, memberId).subscribe(data => {
+        if (data.statusCode != 200) {
+          alert(data.Message);
+        } else {
+          alert("Cancelled Successfully");
+          this.router.navigate(['users', memberId]);
+          this.apiservice.getAppointments(memberId).subscribe(data => {
+            this.appointments = data;
+          })
+        }
       })
-    })
- }
+    } else {
+      alert("Appointment is Safe");
+    }
+
+  }
 }
